feat: puzzle 2 solution
The second puzzle was a bit more complicated presumably because I was
miss calculating the total fuel cost somewhere. I kept getting an answer
that was too high, I think because I double counted some fuel along the
way. This soultion counts it per module and includes the initial module
weight. Thanks for the help Reddit!