feat: second implementation of day 2
I think this is a "better" solution because it reloads the buffers and
doesn't risk storing anything about the program across loop iterations.
I'm still getting the wrong answer according to the side but all the
sample inputs provided work exactly as outlined so I'm not sure where
the problem is.